I added rows to the viability clas boundaries for the script to run. I simply repeated the row for PIX and changed LIT names to have a complete set of LITs.
I modified the graphing script, including a correction to a mistake I had in my original graph script in the WHAPmain_scripts. The original was not corrected yet.

The nes script seems to work and it produces a separate graph for each LIT. Each graph becomes an element of an output list.